+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Support for `all` function in openCypher.
 * User specified transaction execution timeout.
 * Support for query parameters (except for parameters in place of property maps).
+* `degree` function added.
 
 ## v0.6.0